+++ packages/expo-ui/CHANGELOG.md
@@ -40,6 +40,7 @@
 
 - [apple] Migrate remaining `expo-module.config.json` to unified platform syntax. ([#34445](https://github.com/expo/expo/pull/34445) by [@reichhartd](https://github.com/reichhartd))
 - Rename the events for the `Switch` component. ([#34577](https://github.com/expo/expo/pull/34577) by [@behenate](https://github.com/behenate))
+- Allow lower case section titles ([#34931](https://github.com/expo/expo/pull/34931) by [@Pflaumenbaum](https://github.com/Pflaumenbaum))
 
 ## 0.0.1 — 2025-01-21